+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Change log
+## 1.2.0 (09.01.2015)
+1. Added support for UIButton (attributedTitleForSate:).
+2. Added support for UITextField (attributedPlaceholder).
+3. Implemented ability to use NSAttributedString objects as attributed format arguments.
+
 ## 1.1.0 (08.01.2015)
 Implemented ability to update arguments multiple times for the same format.
 This feature is available for UILabel, UITextView and UITextField objects.